src/requirements.txt
apache-airflow
apache-airflow-providers-google
pyarrow
pandas
lxml
html5lib